The script has been rewritten to add support for nanopi neo3 board
