+++ sources/overlays/allwinner/sun7i-a20-analog-codec-disable.dts
@@ -1,0 +1,13 @@
+/dts-v1/;
+/plugin/;
+
+/ {
+    compatible = "allwinner,sun7i-a20";
+
+    fragment@0 {
+	target = <&codec>;
+	__overlay__ {
+	    status = "disabled";
+	};
+    };
+};
