+++ sources/overlays/allwinner/sun7i-a20-spdif-disable.dts
@@ -1,0 +1,13 @@
+/dts-v1/;
+/plugin/;
+
+/ {
+    compatible = "allwinner,sun7i-a20";
+
+    fragment@0 {
+	target = <&spdif>;
+	__overlay__ {
+	    status = "disabled";
+	};
+    };
+};
