--- sources/overlays/sun7i-a20-analog-codec-disable.dts
@@ -1,13 +1,0 @@
-/dts-v1/;
-/plugin/;
-
-/ {
-    compatible = "allwinner,sun7i-a20";
-
-    fragment@0 {
-	target = <&codec>;
-	__overlay__ {
-	    status = "disabled";
-	};
-    };
-};
