--- sources/overlays/sun8i-h3-i2s0-master.dts
@@ -1,49 +1,0 @@
-/dts-v1/;
-/plugin/;
-
-/ {
-	compatible = "allwinner,sun8i-h3";
-
- 	fragment@0 {
- 		target-path = "/";
- 		__overlay__ {
-			pcm5102a: pcm5102a {
-			#sound-dai-cells = <0>;
-			compatible = "ti,pcm5102a";
-			pcm510x,format = "i2s";
-			};
- 		};
- 	};
-
-	fragment@1 {
-		target = <&i2s0>;
-		__overlay__ {
-			status = "okay";
-			pinctrl-0 = <&i2s0_pins>;
-			sound-dai = <&pcm5102a>;
-			pinctrl-names = "default";
-		};
-	};
-
-	fragment@2 {
-		target-path = "/";
-		__overlay__ {
-			sound_i2s {
-				compatible = "simple-audio-card";
-				simple-audio-card,name = "I2S-master";
-				simple-audio-card,mclk-fs = <128>;
-				simple-audio-card,format = "i2s";
-				simple-audio-card,frame-inversion;
-				status = "okay";
-
-				simple-audio-card,cpu {
-					sound-dai = <&i2s0>;
-				};
-
-				simple-audio-card,codec {
-					sound-dai = <&pcm5102a>;
-				};
-			};
-		};
-	};
-};
